requirements.txt
Django==2.2.10

pytz==2019.3

sqlparse==0.3.0

certifi==2019.11.28

chardet==3.0.4

defusedxml==0.6.0

django-ckeditor==5.9.0

django-crispy-forms==1.8.1

django-js-asset==1.2.2

idna==2.9

oauthlib==3.1.0

Pillow==7.0.0

psycopg2-binary==2.8.4

PyJWT==1.7.1

python3-openid==3.1.0

requests==2.23.0

requests-oauthlib==1.3.0

six==1.14.0

social-auth-app-django==3.1.0

social-auth-core==3.2.0
urllib3==1.25.8

django-colorfield==0.2.1

django-colorful==1.3



django-money
django-countries

django-tabbed-changeform-admin
django-easy-select2
paywix
django-admin-shortcuts
django-admin-vali
django-taggit
django-templated-email
django-extra-fields
djangorestframework